Split charts and stuff into their own files and implement certainty interval colors
