+++ src/components/chart/chart.py
@@ -1,0 +1,242 @@
+from abc import ABC, abstractmethod
+from datetime import datetime
+from functools import lru_cache
+from pathlib import Path
+from typing import Any, Self
+
+import pandas as pd
+import plotly.graph_objects as go
+
+from src.components.chart.chart_controls import ChartControls
+from src.components.chart.chart_properties import (
+  Annotation,
+  HorizontalAnnotation,
+  Location,
+  Model,
+  Scenario,
+  VerticalAnnotation,
+  Zoom,
+)
+from src.components.enums import AgeGroup, DataType, Target
+
+BASE_DATA_DIR = Path(__file__).resolve().parent.parent.parent / 'data' / 'rounds'
+
+
+class Chart(ABC):
+  """Abstract base class for charts. Contains common properties and methods for all charts."""
+
+  def __init__(self, controls: ChartControls):
+    """
+    Initialize the base chart class. This gets called first by the concrete chart classes'
+    constructors.
+    """
+    self.controls = controls
+
+    self._raw_df = Chart.collect_data(
+      self.controls.data_type,
+      self.controls.round_num,
+      self.controls.location.name,
+      self.controls.target.input_value,
+    )
+    self._fig: go.Figure = self._create_empty_figure()
+
+  def _create_empty_figure(self) -> go.Figure:
+    return go.Figure()
+
+  @abstractmethod
+  def refresh_fig(self) -> go.Figure:
+    pass
+
+  @abstractmethod
+  def _reload_data(self):
+    pass
+
+  @staticmethod
+  def get_key(controls: ChartControls) -> str:
+    """
+    Generate a unique string key from chart controls.
+    This creates a deterministic string representation that can be used as a dictionary key.
+    """
+    # Create a deterministic string representation of the chart parameters
+    key_components = [
+      str(controls.plot_type),
+      str(controls.round_num),
+      str(sorted([s.name for s in controls.scenarios])),
+      str(sorted([m.name for m in controls.models])),
+      str(controls.location.name),
+      str(controls.target.input_value),
+      str(controls.age_group.input_value),
+      str(controls.certainty_percent.display_value if controls.certainty_percent else ''),
+      str(controls.zoom.x.min if controls.zoom.x.min else ''),
+      str(controls.zoom.x.max if controls.zoom.x.max else ''),
+      str(controls.zoom.y.min if controls.zoom.y.min else ''),
+      str(controls.zoom.y.max if controls.zoom.y.max else ''),
+      str(sorted([a.get_key() for a in controls.annotations]) if controls.annotations else []),
+    ]
+    # Use a deterministic string key
+    return '|'.join(key_components)
+
+  @abstractmethod
+  def __hash__(self) -> int:
+    pass
+
+  @abstractmethod
+  def __eq__(self, other: Self) -> bool:
+    pass
+
+  def get_fig(self) -> go.Figure:
+    return self._fig
+
+  def get_raw_dataframe(self) -> pd.DataFrame:
+    return self._raw_df
+
+  def update_controls(self, controls: ChartControls):
+    """
+    Update the chart controls and refresh the figure without recreating the entire chart object.
+    This method updates the chart parameters and regenerates the figure with new data.
+    """
+    # Only update if the controls have changed
+    if controls != self.controls:
+      self.controls = controls
+      self._reload_data()
+      self.refresh_fig()
+
+  def update_round_num(self, round_num: int):
+    self.controls.round_num = round_num
+    self._reload_data()
+    self.refresh_fig()
+
+  def update_pathogen(self, pathogen: str):
+    self.controls.pathogen = pathogen
+    self._reload_data()
+    self.refresh_fig()
+
+  def update_scenarios(self, scenario_names: list[str]):
+    self.controls.scenarios = [Scenario(name=scenario_name) for scenario_name in scenario_names]
+    self.refresh_fig()
+
+  def update_models(self, model_names: list[str]):
+    self.controls.models = [Model(model_name) for model_name in model_names]
+    self.refresh_fig()
+
+  def update_location(self, location_name: str):
+    self.controls.location = Location(location_name)
+    self._reload_data()
+    self.refresh_fig()
+
+  def update_target(self, target: str):
+    self.controls.target = Target.from_input_value(target)
+    self._reload_data()
+    self.refresh_fig()
+
+  def update_age_group(self, age_group: str):
+    self.controls.age_group = AgeGroup.from_input_value(age_group)
+    self._reload_data()
+    self.refresh_fig()
+
+  def update_annotations(self, annotations: list[dict[str, Any]] | None):
+    self.controls.annotations = (
+      [Annotation.from_dict(annotation) for annotation in annotations] if annotations else None
+    )
+    # might be able to get away with just calling _plot_annotations() here instead of refresh_fig()
+    # since annotations are not data-dependent
+    self.refresh_fig()
+
+  def update_zoom(self, zoom: dict[str, dict[str, Any]] | None):
+    self.controls.zoom = Zoom(zoom['x'], zoom['y']) if zoom else Zoom()
+    self.refresh_fig()
+
+  def _plot_annotations(self):
+    if not self.controls.annotations:
+      return
+    # Clear existing annotations. We will rebuild them below.
+    # self._fig.layout.shapes = []
+    for annotation in self.controls.annotations:
+      if isinstance(annotation, HorizontalAnnotation):
+        self._fig.add_hline(
+          y=annotation.value or 0,
+          line_dash='dot',
+          line_color=annotation.color,
+          line_width=1,
+          annotation_text=annotation.label,
+        )
+      elif isinstance(annotation, VerticalAnnotation):
+        # Convert date to timestamp in milliseconds, as there passing the datetime object directly and adding annotation_text causes a TypeError
+        # See: https://github.com/plotly/plotly.py/issues/3065
+        # Assuming date is in "YYYY-MM-DD" format
+        ms = (
+          datetime.strptime(annotation.value, '%Y-%m-%d').timestamp() * 1000
+          if annotation.value
+          else None
+        )
+
+        self._fig.add_vline(
+          x=ms,
+          line_dash='dot',
+          line_color=annotation.color,
+          line_width=1,
+          annotation_text=annotation.label,
+        )
+      else:
+        raise ValueError(f'Invalid annotation type: {annotation.type}')
+
+  @staticmethod
+  @lru_cache(maxsize=128)
+  def _build_dataset_path(
+    data_type: DataType = DataType.QUANTILE,
+    round_number: int = 1,
+    location_name: str = 'US',
+    target_value: str = 'incident_hospitalization',
+    part: int = 0,
+  ) -> Path:
+    part = str(part)
+    round_fragment = f'round{round_number}'
+    path = (
+      BASE_DATA_DIR
+      / round_fragment
+      / target_value
+      / location_name
+      / data_type.get_path_value()
+      / f'part-{part}.{data_type.get_file_extension()}'
+    )
+    if not path.exists() or not path.is_file():
+      raise FileNotFoundError(f'File not found for chart: {path}')
+    return path
+
+  @staticmethod
+  @lru_cache(maxsize=64)
+  def collect_data(
+    data_type: DataType,
+    round_number: int = 1,
+    location_name: str = 'US',
+    target_value: str = 'incident_hospitalization',
+    part: int = 0,
+  ) -> pd.DataFrame:
+    path = Chart._build_dataset_path(
+      data_type=data_type,
+      round_number=round_number,
+      location_name=location_name,
+      target_value=target_value,
+      part=part,
+    )
+    if not path.exists() or not path.is_file():
+      raise FileNotFoundError(f'File not found for chart: {path}')
+
+    try:
+      return pd.read_csv(path, parse_dates=['target_end_date'])
+    except Exception as e:
+      print(f'Error reading file "{path}": {e}')
+      raise e
+
+  @staticmethod
+  @lru_cache(maxsize=32)
+  def _build_gold_std_path(round_number: int) -> Path:
+    return BASE_DATA_DIR / f'round{round_number}' / 'gold_standard' / 'covid_nhsn_hosp_inc.csv'
+
+  @staticmethod
+  @lru_cache(maxsize=32)
+  def collect_gold_std_data(round_number: int) -> pd.DataFrame:
+    gold_std_path = Chart._build_gold_std_path(round_number)
+    if not gold_std_path.exists() or not gold_std_path.is_file():
+      raise FileNotFoundError(f'File not found for gold standard data: {gold_std_path}')
+    return pd.read_csv(gold_std_path, parse_dates=['time_value'])
