+++ src/components/enums.py
@@ -83,8 +83,10 @@ class DataType(Enum):
 class CertaintyInterval(Enum):
   NONE = ('None', [])
   FIFTY_PERCENT = ('50%', [(0.25, 0.75)])
+  EIGHTY_PERCENT = ('80%', [(0.1, 0.9)])
+  NINETY_PERCENT = ('90%', [(0.05, 0.95)])
   NINETY_FIVE_PERCENT = ('95%', [(0.025, 0.975)])
-  MULTI = ('Multi', [(0.025, 0.975), (0.05, 0.95), (0.1, 0.9), (0.25, 0.75)])
+  ALL = ('All', [(0.025, 0.975), (0.05, 0.95), (0.1, 0.9), (0.25, 0.75)])
 
   def __init__(self, display_value: str, bounds: list[tuple[float, float]]):
     self.display_value = display_value
@@ -97,11 +99,18 @@ class CertaintyInterval(Enum):
     return self.bounds
 
   @classmethod
-  def from_display_value(cls, value: str) -> 'CertaintyInterval':
+  def from_display_value(cls, value: str) -> Self:
     for member in cls:
       if member.display_value == value:
         return member
     raise ValueError(f'No enum member found for display value: {value}')
+
+  @classmethod
+  def from_bounds(cls, bounds: list[tuple[float, float]]) -> Self:
+    for member in cls:
+      if member.bounds == bounds:
+        return member
+    raise ValueError(f'No enum member found for bounds: {bounds}')
 
   @classmethod
   def display_values(cls) -> list[str]:
