--- src/util/slugify.py
@@ -1,1 +1,0 @@
-from slugify import slugify
